pyproject.toml
[project]
name = "GOTTA.BIKE Virtual Team Dicord bot"
version = "0.1.0"
description = "Manage your zwift team on discord. and orginize your racing"
requires-python = ">=3.14"
dependencies = [
    "audioop-lts>=0.2.2",
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "logfire>=4.15.1",
    "py-cord>=2.7.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.8",
]
